Namespace everything as a gem
A bit wordy and annoying, but it's a green build :-D
Still probably needs cleanup, a proper RubyGems page, and updated
install instructions
