+++ Gemfile
@@ -1,16 +1,4 @@
 # frozen_string_literal: true
 
 source "https://rubygems.org"
-
-git_source(:github) {|repo_name| "https://github.com/#{repo_name}" }
-
-gem "wavefile"
-gem "midilib"
-
-group :test do
-	gem "rspec"
-end
-
-group :development, :test do
-	gem "pry"
-end
+gemspec
