+++ lib/inevitable_cacophony.rb
@@ -1,0 +1,2 @@
+module InevitableCacophony
+end
